--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,6 +12,7 @@ exclude = ["/.*", "/tools"]
 description = """
 A simple changelog parser, written in Rust.
 """
+autobenches = false
 
 [package.metadata.docs.rs]
 all-features = true
@@ -19,7 +20,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [workspace]
 resolver = "2"
-members = ["tools/codegen"]
+members = ["benches", "tools/codegen"]
 
 [[bin]]
 name = "parse-changelog"