--- a/Cargo.toml
+++ b/Cargo.toml
@@ -33,6 +33,7 @@ default = ["anyhow", "structopt"]
 # Note: indexmap is public dependency.
 [dependencies]
 indexmap = { version = "1.5.2", features = ["std"] }
+memchr = "2.4"
 once_cell = "1"
 regex = "1"
 
@@ -42,3 +43,9 @@ structopt = { version = "0.3", optional = true }
 [dev-dependencies]
 easy-ext = "0.2"
 tempfile = "3"
+
+[profile.release]
+debug = true
+
+[profile.bench]
+debug = true