This makes parse-changelog more than twice as fast as before.

Before:

```text
parse_changelog         time:   [2.3318 ms 2.3431 ms 2.3574 ms]
```

After:

```text
parse_changelog         time:   [965.72 us 969.73 us 974.69 us]                            
                        change: [-58.853% -58.554% -58.292%] (p = 0.00 < 0.05)
```